--- Framework/Wider.Core.Wpf/ViewModels/ViewAViewModel.cs
@@ -1,25 +1,0 @@
-﻿using Prism.Commands;
-using Prism.Mvvm;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Wider.Shell.ViewModels
-{
-    public class ViewAViewModel : BindableBase
-    {
-        private string _message;
-        public string Message
-        {
-            get { return _message; }
-            set { SetProperty(ref _message, value); }
-        }
-
-        public ViewAViewModel()
-        {
-            Message = "View A from your Prism Module";
-        }
-    }
-}
