+++ Framework/Wider.Core.Wpf/Views/NewFileWindow.xaml
@@ -1,8 +1,8 @@
 ﻿<Window 
-    x:Class="Wider.Shell.Views.NewFileWindow" x:ClassModifier="internal"
+    x:Class="Wider.Core.Views.NewFileWindow" x:ClassModifier="internal"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:Converters="clr-namespace:Wider.Core.Converters;assembly=Wider.Core"
+    xmlns:Converters="clr-namespace:Wider.Core.Converters"
     Title="New Object" Height="400" Width="700"
     WindowStartupLocation="CenterScreen" 
     ResizeMode="NoResize" ShowInTaskbar="False" WindowStyle="ToolWindow">
