+++ Framework/Wider.Core.Wpf/Views/SettingsWindow.xaml
@@ -1,5 +1,5 @@
 ﻿<Window 
-    x:Class="Wider.Shell.Views.SettingsWindow" x:ClassModifier="internal"
+    x:Class="Wider.Core.Views.SettingsWindow" x:ClassModifier="internal"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     Title="Settings" Height="500" Width="700"
