+++ Framework/Wider.Core.Wpf/Wider.Core.Wpf.csproj
@@ -7,8 +7,8 @@
     <ProjectGuid>{F4242FA3-2404-4EA4-95D9-D00057F277D1}</ProjectGuid>
     <OutputType>library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Wider.Shell</RootNamespace>
-    <AssemblyName>Wider.Shell</AssemblyName>
+    <RootNamespace>Wider.Core</RootNamespace>
+    <AssemblyName>Wider.Core</AssemblyName>
     <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <ProjectTypeGuids>{60dc8134-eba5-43b8-bcc9-bb4bc16c2548};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
@@ -35,6 +35,7 @@
   <ItemGroup>
     <Reference Include="System" />
     <Reference Include="System.Data" />
+    <Reference Include="System.Runtime.Serialization" />
     <Reference Include="System.Xml" />
     <Reference Include="Microsoft.CSharp" />
     <Reference Include="System.Core" />
@@ -51,10 +52,7 @@
   <ItemGroup>
     <Compile Include="Converters\DocumentContextMenuMixingConverter.cs" />
     <Compile Include="Service\AbstractSettingsItem.cs" />
-    <Compile Include="Settings\ISettingsManager.cs" />
-    <Compile Include="Settings\SettingsManager.cs" />
     <Compile Include="Service\ToolbarService.cs" />
-    <Compile Include="ShellModule.cs" />
     <Compile Include="Properties\AssemblyInfo.cs">
       <SubType>Code</SubType>
     </Compile>
@@ -68,7 +66,6 @@
       <DependentUpon>Settings.settings</DependentUpon>
       <DesignTimeSharedInput>True</DesignTimeSharedInput>
     </Compile>
-    <Compile Include="Converters\ActiveDocumentConverter.cs" />
     <Compile Include="Unsorted\LayoutInitializer.cs" />
     <Compile Include="Views\NewFileWindow.xaml.cs">
       <DependentUpon>NewFileWindow.xaml</DependentUpon>
@@ -78,11 +75,10 @@
     <Compile Include="Views\SettingsWindow.xaml.cs">
       <DependentUpon>SettingsWindow.xaml</DependentUpon>
     </Compile>
-    <Compile Include="Unsorted\ShellView.xaml.cs">
+    <Compile Include="Views\ShellView.xaml.cs">
       <DependentUpon>ShellView.xaml</DependentUpon>
     </Compile>
     <Compile Include="Unsorted\WideBootstrapper.cs" />
-    <Compile Include="ViewModels\ViewAViewModel.cs" />
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
       <LastGenOutput>Resources.Designer.cs</LastGenOutput>
@@ -94,19 +90,13 @@
     <AppDesigner Include="Properties\" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\Wider.Core\Wider.Core.csproj">
-      <Project>{10205918-47c5-4614-9359-5e30cbe7e289}</Project>
-      <Name>Wider.Core</Name>
-    </ProjectReference>
-  </ItemGroup>
-  <ItemGroup>
     <Page Include="Views\NewFileWindow.xaml">
       <SubType>Designer</SubType>
     </Page>
     <Page Include="Views\SettingsWindow.xaml">
       <SubType>Designer</SubType>
     </Page>
-    <Page Include="Unsorted\ShellView.xaml">
+    <Page Include="Views\ShellView.xaml">
       <SubType>Designer</SubType>
     </Page>
   </ItemGroup>
@@ -116,7 +106,9 @@
     </PackageReference>
     <PackageReference Include="Prism.Autofac" Version="6.3.0" />
   </ItemGroup>
-  <ItemGroup />
+  <ItemGroup>
+    <Folder Include="ViewModels\" />
+  </ItemGroup>
   <Import Project="..\Wider.Core\Wider.Core.projitems" Label="Shared" />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
