+++ Framework/Wider.Core/Wider.Core.projitems
@@ -9,5 +9,74 @@
     <Import_RootNamespace>Wider.Core</Import_RootNamespace>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="$(MSBuildThisFileDirectory)Attributes\FileContentAttribute.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Attributes\NewContentAttribute.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Base\AbstractCommandable.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Base\AbstractPrioritizedTree.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Base\CoreModule.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Base\ICommandable.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Base\IPrioritizedTree.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Base\Workspace.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Controls\AbstractMenuItem.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Controls\AbstractToolbar.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Controls\IToolbar.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Controls\MenuItemViewModel.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Controls\ToolBarItemTemplateSelector.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Controls\ToolbarViewModel.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Controls\WiderStatusbar.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Converters\ActiveDocumentConverter.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Converters\BoolToInsertModeConverter.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Converters\MenuVisibilityConverter.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Converters\NullObjectToVisibiltyConverter.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Converters\NullToBooleanConverter.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Converters\PercentToFontSizeConverter.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Events\ActiveContentChangedEvent.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Events\ClosedContentEvent.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Events\LogEvent.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Events\OpenContentEvent.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Events\SplashCloseEvent.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Events\SplashMessageUpdatedEvent.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Events\ThemeChangeEvent.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Events\WindowClosingEvent.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Services\AbstractWorkspace.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Services\CommandManager.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Services\ContentHandlerRegistry.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Services\ContentModel.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Services\ContentViewModel.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Services\DebugLogService.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Services\ICommandManager.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Services\IContentHandler.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Services\IContentHandlerRegistry.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Services\IContentView.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Services\ILoggerService.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Services\IMenuService.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Services\INewFileWindow.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Services\IOpenDocumentService.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Services\IShell.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Services\IStatusbarService.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Services\ITheme.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Services\IThemeManager.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Services\ITool.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Services\IToolbarService.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Services\IWorkspace.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Services\OpenDocumentService.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Services\PaneLocation.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Services\ThemeManager.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Services\ToolViewModel.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Settings\AbstractSettings.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Settings\IRecentViewItem.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Settings\IRecentViewSettings.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Settings\ISettingsManager.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Settings\IThemeSettings.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Settings\IToolbarPositionSettings.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Settings\IWindowPositionSettings.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Settings\RecentViewItem.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Settings\RecentViewSettings.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Settings\SettingsManager.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Settings\ThemeSettings.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Settings\ToolbarPositionSettings.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Settings\ToolbarSettingItem.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Settings\WindowPositionSettings.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Utils\DatabindingDebugConverter.cs" />
   </ItemGroup>
 </Project>
