+++ Modules/Wider.Content.TextDocument/Wider.Content.TextDocument.csproj
@@ -98,13 +98,13 @@
     </PackageReference>
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\Wider.Core\Wider.Core.csproj">
-      <Project>{10205918-47c5-4614-9359-5e30cbe7e289}</Project>
-      <Name>Wider.Core</Name>
-    </ProjectReference>
+    <Resource Include="Icons\Textfile.png" />
   </ItemGroup>
   <ItemGroup>
-    <Resource Include="Icons\Textfile.png" />
+    <ProjectReference Include="..\..\Framework\Wider.Core.Wpf\Wider.Core.Wpf.csproj">
+      <Project>{f4242fa3-2404-4ea4-95d9-d00057f277d1}</Project>
+      <Name>Wider.Core.Wpf</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
