+++ Modules/Wider.Shell.Metro/Wider.Shell.Metro.csproj
@@ -146,13 +146,9 @@
     <Folder Include="ViewModels\" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\Wider.Core\Wider.Core.csproj">
-      <Project>{10205918-47C5-4614-9359-5E30CBE7E289}</Project>
-      <Name>Wider.Core</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\Wider.Shell\Wider.Shell.csproj">
+    <ProjectReference Include="..\..\Framework\Wider.Core.Wpf\Wider.Core.Wpf.csproj">
       <Project>{f4242fa3-2404-4ea4-95d9-d00057f277d1}</Project>
-      <Name>Wider.Shell</Name>
+      <Name>Wider.Core.Wpf</Name>
     </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
