+++ Wider.sln
@@ -19,13 +19,19 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Wider.Splash", "Modules\Wid
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Wider.Content.TextDocument", "Modules\Wider.Content.TextDocument\Wider.Content.TextDocument.csproj", "{B80E6695-0F27-4AE1-BD7A-EDA42B94A940}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Wider.Core", "Modules\Wider.Core\Wider.Core.csproj", "{10205918-47C5-4614-9359-5E30CBE7E289}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Wider.Shell", "Modules\Wider.Shell\Wider.Shell.csproj", "{F4242FA3-2404-4EA4-95D9-D00057F277D1}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Wider.Shell.Metro", "Modules\Wider.Shell.Metro\Wider.Shell.Metro.csproj", "{53106C03-FE9F-4627-872A-A6E8BCC429FE}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Framework", "Framework", "{34E1BD8E-AEAE-4BFD-A69A-919EC930FC3E}"
+EndProject
+Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "Wider.Core", "Framework\Wider.Core\Wider.Core.shproj", "{6D425467-7F54-4195-93DC-D44EE8CEAD5F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Wider.Core.Wpf", "Framework\Wider.Core.Wpf\Wider.Core.Wpf.csproj", "{F4242FA3-2404-4EA4-95D9-D00057F277D1}"
+EndProject
 Global
+	GlobalSection(SharedMSBuildProjectFiles) = preSolution
+		Framework\Wider.Core\Wider.Core.projitems*{6d425467-7f54-4195-93dc-d44ee8cead5f}*SharedItemsImports = 13
+		Framework\Wider.Core\Wider.Core.projitems*{f4242fa3-2404-4ea4-95d9-d00057f277d1}*SharedItemsImports = 4
+	EndGlobalSection
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
@@ -51,18 +57,14 @@ Global
 		{B80E6695-0F27-4AE1-BD7A-EDA42B94A940}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B80E6695-0F27-4AE1-BD7A-EDA42B94A940}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B80E6695-0F27-4AE1-BD7A-EDA42B94A940}.Release|Any CPU.Build.0 = Release|Any CPU
-		{10205918-47C5-4614-9359-5E30CBE7E289}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{10205918-47C5-4614-9359-5E30CBE7E289}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{10205918-47C5-4614-9359-5E30CBE7E289}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{10205918-47C5-4614-9359-5E30CBE7E289}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F4242FA3-2404-4EA4-95D9-D00057F277D1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F4242FA3-2404-4EA4-95D9-D00057F277D1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F4242FA3-2404-4EA4-95D9-D00057F277D1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F4242FA3-2404-4EA4-95D9-D00057F277D1}.Release|Any CPU.Build.0 = Release|Any CPU
 		{53106C03-FE9F-4627-872A-A6E8BCC429FE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{53106C03-FE9F-4627-872A-A6E8BCC429FE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{53106C03-FE9F-4627-872A-A6E8BCC429FE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{53106C03-FE9F-4627-872A-A6E8BCC429FE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F4242FA3-2404-4EA4-95D9-D00057F277D1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F4242FA3-2404-4EA4-95D9-D00057F277D1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F4242FA3-2404-4EA4-95D9-D00057F277D1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F4242FA3-2404-4EA4-95D9-D00057F277D1}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -73,9 +75,9 @@ Global
 		{625649C0-B1F5-4262-9A38-330FC08A2759} = {F0808592-9E93-4915-9E37-043C56BB9763}
 		{A36AB58C-197B-431B-9DE6-4911DB8F860B} = {92EB10BB-F612-4C23-84BA-95A5D64F851B}
 		{B80E6695-0F27-4AE1-BD7A-EDA42B94A940} = {92EB10BB-F612-4C23-84BA-95A5D64F851B}
-		{10205918-47C5-4614-9359-5E30CBE7E289} = {92EB10BB-F612-4C23-84BA-95A5D64F851B}
-		{F4242FA3-2404-4EA4-95D9-D00057F277D1} = {92EB10BB-F612-4C23-84BA-95A5D64F851B}
 		{53106C03-FE9F-4627-872A-A6E8BCC429FE} = {92EB10BB-F612-4C23-84BA-95A5D64F851B}
+		{6D425467-7F54-4195-93DC-D44EE8CEAD5F} = {34E1BD8E-AEAE-4BFD-A69A-919EC930FC3E}
+		{F4242FA3-2404-4EA4-95D9-D00057F277D1} = {34E1BD8E-AEAE-4BFD-A69A-919EC930FC3E}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {C9AF68FB-8AA0-4C17-A8F9-A88E3EC36423}
